requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
deprecation==2.1.0
fastapi==0.121.0
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
postgrest==2.24.0
propcache==0.4.1
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
realtime==2.24.0
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
